Refactor: Move staking withdrawal helpers to staking subdirectory
